+++ nengo_loihi/tests/test_simulator.py
@@ -1,7 +1,7 @@
 import inspect
 
 import nengo
-from nengo.exceptions import ReadonlyError, ValidationError
+from nengo.exceptions import ReadonlyError, ValidationError, SimulationError
 import numpy as np
 import pytest
 
@@ -11,6 +11,7 @@ from nengo_loihi.builder import Model
 from nengo_loihi.discretize import discretize_model
 from nengo_loihi.emulator import EmulatorInterface
 from nengo_loihi.hardware import HardwareInterface
+from nengo_loihi.hardware.allocators import RoundRobin
 from nengo_loihi.inputs import SpikeInput
 
 
@@ -289,6 +290,22 @@ def test_all_onchip(Simulator):
     assert sim._run_steps.__name__ == "run_steps"
     assert sim.data[out_p].shape[0] == sim.trange().shape[0]
     assert np.all(sim.data[out_p][-1] > 100)
+
+
+@pytest.mark.skipif(pytest.config.getoption('--target') != 'loihi',
+                    reason="snips only on Loihi")
+def test_snips_round_robin_unsupported(Simulator):
+    with nengo.Network() as model:
+        # input is required otherwise precompute will be
+        # automatically overwritten to True (and then no snips)
+        u = nengo.Node(0)
+        x = nengo.Ensemble(1, 1)
+        nengo.Connection(u, x)
+
+    with pytest.raises(SimulationError, match="snips are not supported"):
+        with Simulator(model, precompute=False,
+                       hardware_options={'allocator': RoundRobin(n_chips=8)}):
+            pass
 
 
 def test_progressbar_values(Simulator):